pkg/slack restructure
