+++ deepfence_server/pkg/integration/slack/slack.go
@@ -9,32 +9,6 @@ import (
 // todo: add support for batch size
 const BatchSize = 100
 
-type Slack struct {
-	Config           Config              `json:"config"`
-	IntegrationType  string              `json:"integration_type"`
-	NotificationType string              `json:"notification_type"`
-	Filters          map[string][]string `json:"filters"`
-	Message          string              `json:"message"`
-}
-
-type Config struct {
-	WebhookURL string `json:"webhook_url"`
-	Channel    string `json:"channel"`
-}
-
-type Payload struct {
-	Parse       string `json:"parse,omitempty"`
-	Username    string `json:"username,omitempty"`
-	IconUrl     string `json:"icon_url,omitempty"`
-	IconEmoji   string `json:"icon_emoji,omitempty"`
-	Channel     string `json:"channel,omitempty"`
-	Text        string `json:"text,omitempty"`
-	LinkNames   string `json:"link_names,omitempty"`
-	UnfurlLinks bool   `json:"unfurl_links,omitempty"`
-	UnfurlMedia bool   `json:"unfurl_media,omitempty"`
-	Markdown    bool   `json:"mrkdwn,omitempty"`
-}
-
 func New(b []byte) (*Slack, error) {
 	s := Slack{}
 	err := json.Unmarshal(b, &s)
